@@ -3,40 +3,40 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-brands-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@microsoft/signalr": "^5.0.8",
-    "axios": "^0.21.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-brands-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.16",
+    "@microsoft/signalr": "^5.0.11",
+    "axios": "^0.24.0",
     "bulma": "^0.9.3",
-    "connected-react-router": "^6.9.1",
-    "history": "^4.10.1",
+    "history": "^5.1.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-functional-select": "^3.3.3",
-    "react-redux": "^7.2.4",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-redux": "^7.2.6",
+    "react-router": "^6.0.1",
+    "react-router-dom": "^6.0.1",
     "react-scripts": "^4.0.3",
-    "react-toastify": "^7.0.4",
+    "react-toastify": "^8.1.0",
+    "react-transition-group": "^4.4.2",
     "react-window": "^1.8.6",
-    "redux": "^4.1.0",
-    "redux-thunk": "^2.3.0",
-    "styled-components": "^5.3.0",
-    "web-vitals": "^2.1.0",
-    "workbox-background-sync": "^6.1.5",
-    "workbox-broadcast-update": "^6.1.5",
-    "workbox-cacheable-response": "^6.1.5",
-    "workbox-core": "^6.1.5",
-    "workbox-expiration": "^6.1.5",
-    "workbox-google-analytics": "^6.1.5",
-    "workbox-navigation-preload": "^6.1.5",
-    "workbox-precaching": "^6.1.5",
-    "workbox-range-requests": "^6.1.5",
-    "workbox-routing": "^6.1.5",
-    "workbox-strategies": "^6.1.5",
-    "workbox-streams": "^6.1.5"
+    "redux": "^4.1.2",
+    "redux-thunk": "^2.4.0",
+    "styled-components": "^5.3.3",
+    "web-vitals": "^2.1.2",
+    "workbox-background-sync": "^6.3.0",
+    "workbox-broadcast-update": "^6.3.0",
+    "workbox-cacheable-response": "^6.3.0",
+    "workbox-core": "^6.3.0",
+    "workbox-expiration": "^6.3.0",
+    "workbox-google-analytics": "^6.3.0",
+    "workbox-navigation-preload": "^6.3.0",
+    "workbox-precaching": "^6.3.0",
+    "workbox-range-requests": "^6.3.0",
+    "workbox-routing": "^6.3.0",
+    "workbox-strategies": "^6.3.0",
+    "workbox-streams": "^6.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -61,17 +61,18 @@
   },
   "devDependencies": {
     "@types/history": "^4.7.9",
-    "@types/jest": "^26.0.24",
-    "@types/node": "^16.3.2",
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.18",
-    "@types/react-router": "^5.1.16",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/styled-components": "^5.1.11",
-    "@types/webpack-env": "^1.16.2",
+    "@types/jest": "^27.0.2",
+    "@types/node": "^16.11.7",
+    "@types/react": "^17.0.34",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-redux": "^7.1.20",
+    "@types/react-router": "^5.1.17",
+    "@types/react-router-dom": "^5.3.2",
+    "@types/react-transition-group": "^4.4.4",
+    "@types/styled-components": "^5.1.15",
+    "@types/webpack-env": "^1.16.3",
     "react-hot-loader": "^4.13.0",
-    "sass": "^1.35.2",
-    "typescript": "^4.3.5"
+    "sass": "^1.43.4",
+    "typescript": "^4.4.4"
   }
 }
