@@ -1,4 +1,5 @@
 import { BaseService } from './base.service';
+
 import type { IWeatherForecast } from '../store/weather-forecasts';
 
 /**
@@ -6,18 +7,20 @@ import type { IWeatherForecast } from '../store/weather-forecasts';
  */
 class SampleService extends BaseService {
   private static _sampleService: SampleService;
-  private static _controllerName: string = 'SampleData';
+  private static _controller: string = 'SampleData';
 
-  private constructor(controllerName: string) {
-    super(controllerName);
+  private constructor(name: string) {
+    super(name);
   }
 
   public static get Instance(): SampleService {
-    return this._sampleService || (this._sampleService = new this(this._controllerName));
+    return this._sampleService || (this._sampleService = new this(this._controller));
   }
 
   public async getWeatherForecastsAsync(startDateIndex: number): Promise<IWeatherForecast[]> {
-    const { data } = await this.$http.get<IWeatherForecast[]>(`WeatherForecasts?startDateIndex=${startDateIndex}`);
+    const url = `WeatherForecasts?startDateIndex=${startDateIndex}`;
+    const { data } = await this.$http.get<IWeatherForecast[]>(url);
+
     return data;
   }
 }
