@@ -1,15 +1,18 @@
-import { RoutesConfig } from '../config';
-import { useSelector } from 'react-redux';
-import { NavLink } from 'react-router-dom';
-import { ReactComponent as BulmaLogoSvg } from '../assets/image/BulmaLogo.svg';
+import { useIsLoggedIn } from '../hooks';
+import { Routes as routes } from '../config';
+import { NavLink, generatePath } from 'react-router-dom';
+import { ReactComponent as BulmaLogoSVG } from '../assets/image/BulmaLogo.svg';
 
 import type { Route } from '../config';
-import type { RootState } from '../store';
 import type { FunctionComponent } from 'react';
 
 const Navbar: FunctionComponent = () => {
-  const navRoutes: Route[] = Object.values(RoutesConfig).filter((x) => x.showInNav);
-  const isAuthenticated = useSelector<RootState, boolean>(state => state.auth.isAuthenticated);
+  const isLoggedIn = useIsLoggedIn();
+
+  const navRoutes = routes.reduce((acc: Route[], r: Route) => {
+    r.showInNav && acc.push(r);
+    return acc;
+  }, []);
 
   return (
     <nav
@@ -19,24 +22,22 @@ const Navbar: FunctionComponent = () => {
     >
       <div className='navbar-wrapper'>
         <div className='brand-wrapper'>
-          <BulmaLogoSvg
-            width='135'
-            height='66'
+          <BulmaLogoSVG
+            width='130'
+            height='65'
             aria-hidden={true}
             title='bulma.io-logo'
           />
         </div>
         <div className='navbar-routes'>
-          {isAuthenticated &&
-            navRoutes.map(({ path, exact, displayName }) => (
+          {isLoggedIn &&
+            navRoutes.map(({ path, name, params }) => (
               <NavLink
-                to={path}
-                key={path}
-                exact={exact}
-                className='navbar-item'
-                activeClassName='is-active'
+                key={name}
+                to={generatePath(path, params)}
+                className={({ isActive }) => 'navbar-item' + (isActive ? ' is-active' : '')}
               >
-                {displayName}
+                {name}
               </NavLink>
             ))}
         </div>
