@@ -6,29 +6,32 @@ import type { RootState } from '../../store';
 import type { FunctionComponent } from 'react';
 
 const CounterFormGroup: FunctionComponent = () => {
-  const dispatch = useDispatch();
   const count = useSelector<RootState, number>((state) => state.form.count);
+
+  const dispatch = useDispatch();
+  const decrement = () => dispatch(actionCreators.decrement());
+  const increment = () => dispatch(actionCreators.increment());
 
   return (
     <div className='column'>
       <h3 className='title is-4'>Counter</h3>
-      <h5 className='subtitle is-5'>Use buttons to update count value</h5>
+      <h5 className='subtitle is-5'>Use buttons to increment/decrement</h5>
       <p className='buttons incrementer-buttons form-control-group'>
         <button
-          className='button is-light minus'
-          onClick={() => dispatch(actionCreators.decrement())}
+          onClick={decrement}
+          className='button is-info'
         >
           <FontAwesomeIcon icon='minus' />
         </button>
         <button
-          className='button is-light plus'
-          onClick={() => dispatch(actionCreators.increment())}
+          onClick={increment}
+          className='button is-info'
         >
           <FontAwesomeIcon icon='plus' />
         </button>
       </p>
       <p className='subtitle is-5'>
-        Current count: <code>{count}</code>
+        Value: <code className="form-value">{count}</code>
       </p>
     </div>
   );
