@@ -1,7 +1,3 @@
-import useTextInput from './useTextInput';
-import useOnClickOutside from './useOnClickOutside';
-
-export {
-  useTextInput,
-  useOnClickOutside
-};
+export { default as useTextInput } from './useTextInput';
+export { default as useIsLoggedIn } from './useIsLoggedIn';
+export { default as useOnClickOutside } from './useOnClickOutside';
