@@ -2,4 +2,9 @@ export { reducer } from './reducer';
 export { actionCreators } from './actions';
 export { AuthActionType, AuthStatusEnum } from './types';
 
-export type { AuthPayload, IAuthUser, IAuthState, ICredentials } from './types';
+export type {
+  AuthPayload,
+  IAuthUser,
+  IAuthState,
+  ICredentials
+} from './types';
