@@ -8,29 +8,30 @@ export const actionCreators = {
   resetState: (): ReduxAction<WeatherForecastPayload> => ({
     type: WeatherActionType.RESET_STATE,
   }),
-  requestWeatherForecasts: (
-    startDateIndex: number
-  ): AppThunk<WeatherForecastPayload> => async (dispatch, getState) => {
-    // If param startDateIndex === state.startDateIndex, do not perform action
-    if (startDateIndex === getState().weatherForecasts.startDateIndex) {
-      return;
-    }
+  requestWeatherForecasts:
+    (startDateIndex: number): AppThunk<WeatherForecastPayload> =>
+    async (dispatch, getState) => {
+      // If param startDateIndex === state.startDateIndex, do not perform action
+      if (startDateIndex === getState().weatherForecasts.startDateIndex) {
+        return;
+      }
 
-    // Dispatch request to intialize loading phase
-    dispatch({
-      payload: { startDateIndex },
-      type: WeatherActionType.REQUEST
-    });
-
-    // Build http request and success handler in Promise<void> wrapper / complete processing
-    try {
-      const forecasts = await SampleApi.getWeatherForecastsAsync(startDateIndex);
+      // Dispatch request to intialize loading phase
       dispatch({
-        type: WeatherActionType.RECEIVE,
-        payload: { forecasts, startDateIndex }
+        payload: { startDateIndex },
+        type: WeatherActionType.REQUEST,
       });
-    } catch (e) {
-      console.error(e);
-    }
-  },
+
+      // Build http request and success handler in Promise<void> wrapper / complete processing
+      try {
+        const forecasts = await SampleApi.getWeatherForecastsAsync(startDateIndex);
+
+        dispatch({
+          type: WeatherActionType.RECEIVE,
+          payload: { forecasts, startDateIndex },
+        });
+      } catch (e) {
+        console.error(e);
+      }
+    },
 };
