@@ -2,4 +2,8 @@ export { reducer } from './reducer';
 export { actionCreators } from './actions';
 export { WeatherActionType } from './types';
 
-export type { IWeatherForecast, IWeatherForecastsState, WeatherForecastPayload } from './types';
+export type {
+  IWeatherForecast,
+  IWeatherForecastsState,
+  WeatherForecastPayload
+} from './types';
