@@ -1,3 +1,1 @@
-const classNames = (classes: any[]): string => classes.filter(Boolean).join(' ');
-
-export default classNames;
+export const classNames = (cls: any[]): string => cls.filter(Boolean).join(' ');
